# ainovex-admin-and-website/admin-panel/lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AINOVEX administration and secure GitHub publishing API
 * OpenAPI spec version: 0.2.0
 */

export interface ErrorResponse {
  error: string;
}